# shared/schemas/api.openapi.json
{
  "components": {
    "schemas": {
...
      "AutoFix": {
        "additionalProperties": true,
        "description": "Identifier + payload for a single auto-applied fix.\n\nThe list of supported fix types lives in docs/JJSCAN_PLUS_RULES.md and is\nenforced by the JJSCAN+ Bob implementation.",
        "properties": {
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "fix": {
            "title": "Fix",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "fix"
        ],
        "title": "AutoFix",
        "type": "object"
      },
...
      "Envelope_PromoteCancelResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PromoteCancelResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[PromoteCancelResponse]",
        "type": "object"
      },
      "Envelope_PromoteResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PromoteResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[PromoteResponse]",
        "type": "object"
      },
      "Envelope_QueueDecisionResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueueDecisionResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[QueueDecisionResponse]",
        "type": "object"
      },
      "Envelope_QueueListResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueueListResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[QueueListResponse]",
        "type": "object"
      },
      "Envelope_QueueSinceResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QueueSinceResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[QueueSinceResponse]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromoteCancelResponse": {
        "properties": {
          "promote_event_id": {
            "title": "Promote Event Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ReviewState"
          }
        },
        "required": [
          "promote_event_id",
          "state"
        ],
        "title": "PromoteCancelResponse",
        "type": "object"
      },
      "PromoteRequest": {
        "properties": {
          "auto_apply_fixes": {
            "items": {
              "type": "string"
            },
            "title": "Auto Apply Fixes",
            "type": "array"
          },
          "jcl_name": {
            "title": "Jcl Name",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source_region": {
            "title": "Source Region",
            "type": "string"
          },
          "target_region": {
            "title": "Target Region",
            "type": "string"
          }
        },
        "required": [
          "jcl_name",
          "source_region",
          "target_region"
        ],
        "title": "PromoteRequest",
        "type": "object"
      },
      "PromoteResponse": {
        "properties": {
          "auto_fixes_applied": {
            "items": {
              "$ref": "#/components/schemas/AutoFix"
            },
            "title": "Auto Fixes Applied",
            "type": "array"
          },
          "auto_fixes_available_but_not_applied": {
            "items": {
              "$ref": "#/components/schemas/AutoFix"
            },
            "title": "Auto Fixes Available But Not Applied",
            "type": "array"
          },
          "confidence_breakdown": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Confidence Breakdown",
            "type": "object"
          },
          "confidence_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "type": "integer"
          },
          "diff": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Diff",
            "type": "array"
          },
          "promote_event_id": {
            "title": "Promote Event Id",
            "type": "string"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          },
          "state": {
            "$ref": "#/components/schemas/ReviewState"
          }
        },
        "required": [
          "promote_event_id",
          "diff",
          "confidence_score",
          "auto_fixes_applied",
          "state"
        ],
        "title": "PromoteResponse",
        "type": "object"
      },
...
      "QueueDecisionRequest": {
        "properties": {
          "reason": {
            "maxLength": 400,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "QueueDecisionRequest",
        "type": "object"
      },
      "QueueDecisionResponse": {
        "properties": {
          "audit_event_id": {
            "title": "Audit Event Id",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decided_by": {
            "title": "Decided By",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ReviewState"
          }
        },
        "required": [
          "event_id",
          "state",
          "decided_by",
          "decided_at",
          "audit_event_id"
        ],
        "title": "QueueDecisionResponse",
        "type": "object"
      },
      "QueueItem": {
        "additionalProperties": true,
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "initiator": {
            "title": "Initiator",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ReviewState"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "type",
          "initiator",
          "ts",
          "state"
        ],
        "title": "QueueItem",
        "type": "object"
      },
      "QueueListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "last_seq": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seq"
          },
          "pending_count": {
            "title": "Pending Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "pending_count"
        ],
        "title": "QueueListResponse",
        "type": "object"
      },
      "QueueSinceResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "last_seq": {
            "title": "Last Seq",
            "type": "string"
          }
        },
        "required": [
          "items",
          "last_seq"
        ],
        "title": "QueueSinceResponse",
        "type": "object"
      },
...
      "ReviewState": {
        "enum": [
          "draft",
          "pending_review",
          "approved",
          "rejected",
          "cancelled"
        ],
        "title": "ReviewState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI control plane for z/OS modernization. See docs/API.md.",
    "title": "Helios Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/promote": {
      "post": {
        "operationId": "promote_api_promote_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PromoteResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote",
        "tags": [
          "promote"
        ]
      }
    },
    "/api/promote/{event_id}": {
      "get": {
        "operationId": "get_promote_api_promote__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PromoteResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Promote",
        "tags": [
          "promote"
        ]
      }
    },
    "/api/promote/{event_id}/cancel": {
      "post": {
        "operationId": "cancel_promote_api_promote__event_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PromoteCancelResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Promote",
        "tags": [
          "promote"
        ]
      }
    },
    "/api/queue": {
      "get": {
        "operationId": "list_queue_api_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewState"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending_review",
              "title": "State"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Queue",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/queue/since": {
      "get": {
        "operationId": "since_api_queue_since_get",
        "parameters": [
          {
            "in": "query",
            "name": "seq",
            "required": true,
            "schema": {
              "title": "Seq",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueSinceResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Since",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/queue/{event_id}/approve": {
      "post": {
        "operationId": "approve_api_queue__event_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueDecisionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "queue"
        ]
      }
    },
    "/api/queue/{event_id}/reject": {
      "post": {
        "operationId": "reject_api_queue__event_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_QueueDecisionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject",
        "tags": [
          "queue"
        ]
      }
    },
...
  }
}